+++ CHANGELOG.md
@@ -2,9 +2,8 @@
 
 ## v1.0.0-rc3 (unreleased)
 
-- **[Breaking]** Rename `fx.Inject` to `fx.Extract` to avoid using the
-  term "inject" outside of the ideal usecase. `fx.Extract` is a hammer to
-  break glass, not a way to achieve pure dependency injection.
+- **[Breaking]** Rename `fx.Inject` to `fx.Extract`.
+- `fx.Extract` now supports `fx.In` tags on target structs.
 
 ## v1.0.0-rc2 (21 Jul 2017)